CallID class
