+++ src/batchMultiSigCall/classes/ImportFCT/index.ts
@@ -1,7 +1,7 @@
 import { ChainId } from "@kirobo/ki-eth-fct-provider-ts";
-import { parseSessionID } from "batchMultiSigCall/helpers";
-import _ from "lodash";
 import { IBatchMultiSigCallFCT, RequiredFCTOptions, TypedDataDomain } from "types";
+
+import { SessionID } from "../SessionID";
 
 export class ImportFCT {
   public options: RequiredFCTOptions;
@@ -14,9 +14,8 @@ export class ImportFCT {
     const { meta } = FCT.typedData.message;
     const domain = FCT.typedData.domain;
 
-    const sessionIdOptions = parseSessionID(FCT.sessionId, FCT.builder, FCT.externalSigners);
-    const name = meta.name;
-    this.options = _.merge({}, sessionIdOptions, { name });
+    const sessionIdOptions = SessionID.fromFCT(FCT);
+    this.options = sessionIdOptions;
     this.randomId = meta.random_id;
     this.version = meta.version;
     this.chainId = domain.chainId.toString() as ChainId;
