+++ src/batchMultiSigCall/classes/index.ts
@@ -1,3 +1,4 @@
+export * from "./CallID";
 export * from "./EIP712";
 export * from "./EIP712StructTypes";
 export * from "./ExportFCT";
